# static/swagger-ui/swagger/openapi.yaml
swagger: "2.0"
info:
  title: JeeekAPI
  description: エンジニアのための自己実現SNSアプリ
  contact:
    name: tonouchi510
    email: tonouchi27@gmail.com
    url: https://github.com/tonouchi510/Jeeek/issues
  version: "0.1"
host: localhost:8080
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /v1/admin/signin:
    post:
      tags:
      - Admin
      summary: admin signin Admin
      description: admin権限のトークンを取得します．
      operationId: Admin#admin signin
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: Admin SigninRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/AdminAdminSigninRequestBody'
          required:
          - uid
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/AdminAdminSigninResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/AdminAdminSigninUnauthorizedResponseBody'
      schemes:
      - http
  /v1/admin/users:
    get:
      tags:
      - Admin
      summary: admin list user Admin
      description: ユーザーの一覧を返します。
      operationId: Admin#admin list user
      parameters:
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/AdminJeeekUserResponseCollection'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/AdminAdminListUserUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
    post:
      tags:
      - Admin
      summary: admin create new user Admin
      description: 新しいユーザーを登録します。
      operationId: Admin#admin create new user
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      - name: Admin Create New UserRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/AdminAdminCreateNewUserRequestBody'
          required:
          - user_name
          - email_address
          - phone_number
          - photo_url
      responses:
        "201":
          description: Created response.
          schema:
            $ref: '#/definitions/AdminAdminCreateNewUserResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/AdminAdminCreateNewUserUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
  /v1/admin/users/{user_id}:
    get:
      tags:
      - Admin
      summary: admin get user Admin
      description: 指定したIDのユーザーの情報を返します。
      operationId: Admin#admin get user
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: user_id
        in: path
        description: User id of firebase
        required: true
        type: string
        maxLength: 28
        minLength: 28
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/AdminAdminGetUserResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/AdminAdminGetUserUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
    put:
      tags:
      - Admin
      summary: admin update user Admin
      description: 指定したユーザー情報を更新します。
      operationId: Admin#admin update user
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: user_id
        in: path
        description: User id of firebase
        required: true
        type: string
        maxLength: 28
        minLength: 28
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      - name: Admin Update UserRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/AdminAdminUpdateUserRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/AdminAdminUpdateUserResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/AdminAdminUpdateUserUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
    delete:
      tags:
      - Admin
      summary: admin delete user Admin
      description: 指定したユーザーを削除します。
      operationId: Admin#admin delete user
      parameters:
      - name: user_id
        in: path
        description: User id of firebase
        required: true
        type: string
        maxLength: 28
        minLength: 28
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      responses:
        "204":
          description: No Content response.
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/AdminAdminDeleteUserUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
  /v1/users:
    get:
      tags:
      - User
      summary: List user User
      description: ユーザーの一覧を返します。
      operationId: User#List user
      parameters:
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserJeeekUserResponseCollection'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/UserListUserUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
    put:
      tags:
      - User
      summary: Update user User
      description: 現在のセッションに紐付くユーザー情報を更新します。
      operationId: User#Update user
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      - name: Update UserRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/UserUpdateUserRequestBody'
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserUpdateUserResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/UserUpdateUserUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
    delete:
      tags:
      - User
      summary: Delete user User
      description: 現在のセッションに紐づくユーザーを削除します。
      operationId: User#Delete user
      parameters:
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      responses:
        "204":
          description: No Content response.
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/UserDeleteUserUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
  /v1/users/{user_id}:
    get:
      tags:
      - User
      summary: Get user User
      description: 指定したIDのユーザーの情報を返します。
      operationId: User#Get user
      produces:
      - application/json; charset=utf-8
      parameters:
      - name: user_id
        in: path
        description: User id of firebase
        required: true
        type: string
        maxLength: 28
        minLength: 28
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/UserGetUserResponseBody'
        "401":
          description: Unauthorized response.
          schema:
            $ref: '#/definitions/UserGetUserUnauthorizedResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
...
definitions:
...
  AdminAdminCreateNewUserRequestBody:
    title: AdminAdminCreateNewUserRequestBody
    type: object
    properties:
      email_address:
        type: string
        description: ユーザーのプライマリ メールアドレス
        example: keisuke.honda+testuser@ynu.jp
        format: email
      phone_number:
        type: string
        description: ユーザのメインの電話番号
        example: "08079469367"
        pattern: ^\+?[\d]{10,}$
      photo_url:
        type: string
        description: ユーザーの写真 URL
        example: https://imageurl.com
      user_name:
        type: string
        description: ユーザーの表示名
        example: keisuke.honda
        minLength: 1
        maxLength: 20
    example:
      email_address: keisuke.honda+testuser@ynu.jp
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_name: keisuke.honda
    required:
    - user_name
    - email_address
    - phone_number
    - photo_url
  AdminAdminCreateNewUserResponseBody:
    title: 'Mediatype identifier: application/vnd.jeeek.user+json; view=default'
    type: object
    properties:
      email_address:
        type: string
        description: ユーザーのプライマリ メールアドレス
        example: keisuke.honda+testuser@ynu.jp
        format: email
      email_verified:
        type: boolean
        description: ユーザーのプライマリ メールアドレスが確認されているかどうか
        example: false
      phone_number:
        type: string
        description: ユーザのメインの電話番号
        example: "08079469367"
        pattern: ^\+?[\d]{10,}$
      photo_url:
        type: string
        description: ユーザーの写真 URL
        example: https://imageurl.com
      user_id:
        type: string
        description: User id of firebase
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      user_name:
        type: string
        description: ユーザーの表示名
        example: keisuke.honda
        minLength: 1
        maxLength: 20
    description: Admin Create New UserResponseBody result type (default view)
    example:
      email_address: keisuke.honda+testuser@ynu.jp
      email_verified: false
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_id: XRQ85mtXnINISH25zfM0m5RlC6L2
      user_name: keisuke.honda
    required:
    - user_id
    - user_name
    - email_address
    - photo_url
    - phone_number
    - email_verified
  AdminAdminCreateNewUserUnauthorizedResponseBody:
    title: AdminAdminCreateNewUserUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: Consequuntur est magnam.
  AdminAdminDeleteUserUnauthorizedResponseBody:
    title: AdminAdminDeleteUserUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: Et expedita.
  AdminAdminGetUserResponseBody:
    title: 'Mediatype identifier: application/vnd.jeeek.user+json; view=default'
    type: object
    properties:
      email_address:
        type: string
        description: ユーザーのプライマリ メールアドレス
        example: keisuke.honda+testuser@ynu.jp
        format: email
      email_verified:
        type: boolean
        description: ユーザーのプライマリ メールアドレスが確認されているかどうか
        example: true
      phone_number:
        type: string
        description: ユーザのメインの電話番号
        example: "08079469367"
        pattern: ^\+?[\d]{10,}$
      photo_url:
        type: string
        description: ユーザーの写真 URL
        example: https://imageurl.com
      user_id:
        type: string
        description: User id of firebase
        example: XRQ85mtXnINISH25zfM0m5RlC6L2
        minLength: 28
        maxLength: 28
      user_name:
        type: string
        description: ユーザーの表示名
        example: keisuke.honda
        minLength: 1
        maxLength: 20
    description: Admin Get UserResponseBody result type (default view)
    example:
      email_address: keisuke.honda+testuser@ynu.jp
      email_verified: false
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_id: XRQ85mtXnINISH25zfM0m5RlC6L2
      user_name: keisuke.honda
    required:
    - user_id
    - user_name
    - email_address
    - photo_url
    - phone_number
    - email_verified
  AdminAdminGetUserUnauthorizedResponseBody:
    title: AdminAdminGetUserUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: Rem accusantium ut et voluptas est.
...
  AdminAdminListUserUnauthorizedResponseBody:
    title: AdminAdminListUserUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: Perspiciatis magni laboriosam.
  AdminAdminSigninRequestBody:
    title: AdminAdminSigninRequestBody
    type: object
    properties:
      uid:
        type: string
        example: 4WIbqiNIpIYXqrfBMVZsbKCepau1
    example:
      uid: 4WIbqiNIpIYXqrfBMVZsbKCepau1
    required:
    - uid
  AdminAdminSigninResponseBody:
    title: 'Mediatype identifier: application/vnd.jeeek.admin.signin+json; view=default'
    type: object
    properties:
      token:
        type: string
        description: JWT used for authentication
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
    description: Admin SigninResponseBody result type (default view)
    example:
      token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
    required:
    - token
  AdminAdminSigninUnauthorizedResponseBody:
    title: AdminAdminSigninUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: Cum modi.
  AdminAdminUpdateUserRequestBody:
    title: AdminAdminUpdateUserRequestBody
    type: object
    properties:
      disabled:
        type: boolean
        description: ユーザが停止状態かどうか（論理削除）
        example: true
      email_address:
        type: string
        description: ユーザーのプライマリ メールアドレス
        example: keisuke.honda+testuser@ynu.jp
        format: email
      email_verified:
        type: boolean
        description: ユーザーのプライマリ メールアドレスが確認されているかどうか
        example: false
      phone_number:
        type: string
        description: ユーザのメインの電話番号
        example: "08079469367"
        pattern: ^\+?[\d]{10,}$
      photo_url:
        type: string
        description: ユーザーの写真 URL
        example: https://imageurl.com
      user_name:
        type: string
        description: ユーザーの表示名
        example: keisuke.honda
        minLength: 1
        maxLength: 20
    example:
      disabled: false
      email_address: keisuke.honda+testuser@ynu.jp
      email_verified: true
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_name: keisuke.honda
...
  AdminAdminUpdateUserUnauthorizedResponseBody:
    title: AdminAdminUpdateUserUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: Molestiae id itaque voluptatem.
  AdminJeeekUserResponseCollection:
    title: 'Mediatype identifier: application/vnd.jeeek.user+json; type=collection;
      view=default'
    type: array
    items:
      $ref: '#/definitions/JeeekUserResponse'
    description: Admin List UserResponseBody is the result type for an array of JeeekUserResponse
      (default view)
    example:
    - email_address: keisuke.honda+testuser@ynu.jp
      email_verified: false
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_id: XRQ85mtXnINISH25zfM0m5RlC6L2
      user_name: keisuke.honda
    - email_address: keisuke.honda+testuser@ynu.jp
      email_verified: false
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_id: XRQ85mtXnINISH25zfM0m5RlC6L2
      user_name: keisuke.honda
    - email_address: keisuke.honda+testuser@ynu.jp
      email_verified: false
      phone_number: "08079469367"
      photo_url: https://imageurl.com
      user_id: XRQ85mtXnINISH25zfM0m5RlC6L2
      user_name: keisuke.honda
...
  UserDeleteUserUnauthorizedResponseBody:
    title: UserDeleteUserUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: A architecto.
...
  UserListUserUnauthorizedResponseBody:
    title: UserListUserUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: Enim consectetur a qui quia dicta.
...
  UserUpdateUserUnauthorizedResponseBody:
    title: UserUpdateUserUnauthorizedResponseBody
    type: string
    description: Credentials are invalid
    example: Eaque molestiae omnis eaque consequatur rerum cum.
securityDefinitions:
  jwt_header_Authorization:
    type: apiKey
    description: Secures endpoint by requiring a valid JWT token retrieved via the
      firebase.
    name: Authorization
    in: header
...